loader: distinquish between unknown and no command
Added a new config option to CommandLoader to set the command to run when
no command was specified on the command line.

This new option allows class consumers to distiquish between the case when
an unknown command was requested (slc wants to run the script of that name)
and the case where no command was specified at all (slc want to print usage).

This commit also fixes the problem where running `slc` without any parameters
in the slc source directory caused an infinite loop: running `slc` loaded
slc module into REPL; loading of slc module started a new REPL and loaded
slc module again again, and so on.

See issue #27.