slc update command
npm install -g will install peerDependencies, if they don't exist, but
it won't update them, it only complains about them. Add an slc update
command to update strong-cli and its dependencies. It will fail and dump
error if at the end, their are peerDependency problems. Unfortunately,
due to a mis-design of npm, any globally installed module with a problem
will cause strong-cli to appear to have a dependency problem, but we
can't help that.