Partially revert 7479d2250: Fix windows spawn
That commit fixed windows, but changed how slc would call `cmd arg1
arg2` on unix to:
  /bin/sh -c cmd arg1 arg2
the effect of which is that cmd would be executed with no arguments. slc
run, slc debug, etc., would all appear to work, unless arguments were
supplied, in which case they would not be passed to the cmd.