strongops: remove --register and cleanup options
--register is deprecated as part of the work flow, see SLN-810. Login is
now the only option.

--colors/--no-colors is silly, there is no point in a colorize option
for a one shot command when it's off by default.

--saveall is a debug option, it should never be used by customers (it
exists, still, but is undocumented).

I'm concerned that the decision on --register will be reversed, so much
of the code is still there. We are also planning on refactoring this
code into strong-agent, at which point it will be stripped to a minimum,
I don't want to that now.