command: always set CMD so children can use it
lib/loader.js sets env.SLC_COMMAND, but then all children need to
prepend "slc " to this get the original command name (used in usage
messages). node-debug uses the less slc-specific CMD. With this change,
CMD is set to the full "slc whatever" for all sub-commands, so
node-debug is less of a special case.