strongops: Add --license option for saving license
When the --license option is specified no authentication is done, but
the given license key is validated and then written to strongloop.json
if it is valid.

If a user also needs to use StrongOps, they can run the command without
the --license option to fill in the authentiction details.

Adds dependency on not-yet-releases strong-agent@0.4.13

SLN-1223 #refs