strong-cli: Replace README with deprecation notice
To be published as a patch release of the old strong-cli package so
that package page on https://www.npmjs.com/package/strong-cli directs
people to install strongloop instead (or upgrade if already installed).

Additionally, we will add a deprecation notice to the strong-cli
package so that anyone installing it will see a warning in their
console if they install strong-cli.